Refactoring of file structure; setting up setuptools package
