+++ .gitignore
@@ -1,2 +1,5 @@
 __pycache__
 .pyc
+.DS_Store
+.ropeproject
+dist
