--- Compute_RSCU_tw/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Thu Aug 12 10:01:48 2021
-
-@author: RWint
-"""
-
