+++ src/codonpy.egg-info/SOURCES.txt
@@ -1,0 +1,20 @@
+README.md
+pyproject.toml
+src/codonpy/__init__.py
+src/codonpy/aa_usage.py
+src/codonpy/fasta2csv.py
+src/codonpy/fix_fasta.py
+src/codonpy.egg-info/PKG-INFO
+src/codonpy.egg-info/SOURCES.txt
+src/codonpy.egg-info/dependency_links.txt
+src/codonpy.egg-info/entry_points.txt
+src/codonpy.egg-info/requires.txt
+src/codonpy.egg-info/top_level.txt
+src/codonpy/CodonCount/__init__.py
+src/codonpy/CodonCount/__main__.py
+src/codonpy/CodonUsage_per_1000/__init__.py
+src/codonpy/CodonUsage_per_1000/__main__.py
+src/codonpy/Compute_RSCU_gene/__init__.py
+src/codonpy/Compute_RSCU_gene/__main__.py
+src/codonpy/Compute_RSCU_tw/__init__.py
+src/codonpy/Compute_RSCU_tw/__main__.py
