+++ src/codonpy.egg-info/requires.txt
@@ -1,0 +1,2 @@
+pandas
+numpy
