+++ src/codonpy/fas2csv.sh
