+++ src/codonpy/test_data/NB_CDS5_fixed.fasta
