+++ src/codonpy/test_data/results_tw.rscu
