--- test_data/df_count2.csv
@@ -1,60 +1,0 @@
-Codon,Obs_Freq,Amino_acid,Length,SeqID
-UUU,10,Phe,2166,>FBpp0423064
-UUC,8,Phe,2166,>FBpp0423064
-UCU,8,Ser4,2166,>FBpp0423064
-UCC,10,Ser4,2166,>FBpp0423064
-UCA,11,Ser4,2166,>FBpp0423064
-UCG,11,Ser4,2166,>FBpp0423064
-AGU,17,Ser2,2166,>FBpp0423064
-AGC,13,Ser2,2166,>FBpp0423064
-CUU,6,Leu4,2166,>FBpp0423064
-CUC,7,Leu4,2166,>FBpp0423064
-CUA,10,Leu4,2166,>FBpp0423064
-CUG,7,Leu4,2166,>FBpp0423064
-UUA,7,Leu2,2166,>FBpp0423064
-UUG,19,Leu2,2166,>FBpp0423064
-UAU,11,Tyr,2166,>FBpp0423064
-UAC,3,Tyr,2166,>FBpp0423064
-UGU,7,Cys,2166,>FBpp0423064
-UGC,3,Cys,2166,>FBpp0423064
-CGU,4,Arg4,2166,>FBpp0423064
-CGC,6,Arg4,2166,>FBpp0423064
-CGA,5,Arg4,2166,>FBpp0423064
-CGG,4,Arg4,2166,>FBpp0423064
-AGA,13,Arg2,2166,>FBpp0423064
-AGG,9,Arg2,2166,>FBpp0423064
-CCU,7,Pro,2166,>FBpp0423064
-CCC,14,Pro,2166,>FBpp0423064
-CCA,22,Pro,2166,>FBpp0423064
-CCG,7,Pro,2166,>FBpp0423064
-CAU,7,His,2166,>FBpp0423064
-CAC,4,His,2166,>FBpp0423064
-CAA,16,Gln,2166,>FBpp0423064
-CAG,18,Gln,2166,>FBpp0423064
-AUU,14,Ile,2166,>FBpp0423064
-AUC,13,Ile,2166,>FBpp0423064
-AUA,15,Ile,2166,>FBpp0423064
-ACU,9,Thr,2166,>FBpp0423064
-ACC,10,Thr,2166,>FBpp0423064
-ACA,9,Thr,2166,>FBpp0423064
-ACG,6,Thr,2166,>FBpp0423064
-AAU,31,Asn,2166,>FBpp0423064
-AAC,14,Asn,2166,>FBpp0423064
-AAA,41,Lys,2166,>FBpp0423064
-AAG,38,Lys,2166,>FBpp0423064
-GUU,11,Val,2166,>FBpp0423064
-GUC,12,Val,2166,>FBpp0423064
-GUA,7,Val,2166,>FBpp0423064
-GUG,8,Val,2166,>FBpp0423064
-GCU,11,Ala,2166,>FBpp0423064
-GCC,12,Ala,2166,>FBpp0423064
-GCA,8,Ala,2166,>FBpp0423064
-GCG,4,Ala,2166,>FBpp0423064
-GAU,31,Asp,2166,>FBpp0423064
-GAC,7,Asp,2166,>FBpp0423064
-GAA,36,Glu,2166,>FBpp0423064
-GAG,18,Glu,2166,>FBpp0423064
-GGU,9,Gly,2166,>FBpp0423064
-GGC,7,Gly,2166,>FBpp0423064
-GGA,10,Gly,2166,>FBpp0423064
-GGG,2,Gly,2166,>FBpp0423064
